Reorganise pixel driller layer files
